# docs/swagger.yaml
basePath: /v1
definitions:
...
  dtos.Celebration:
    type: object
  dtos.Content:
    properties:
      celebration:
        $ref: '#/definitions/dtos.Celebration'
      celebrations:
        items:
          $ref: '#/definitions/dtos.Celebration'
        type: array
      post:
        $ref: '#/definitions/dtos.Post'
      title:
        type: string
      type:
        type: string
    type: object
  dtos.PagedContent:
    properties:
      contnet:
        items:
          $ref: '#/definitions/dtos.Content'
        type: array
      limit:
        type: integer
      page:
        type: integer
    type: object
  dtos.PagedPosts:
    properties:
      limit:
        type: integer
      page:
        type: integer
      posts:
        items:
          $ref: '#/definitions/dtos.Post'
        type: array
    type: object
...
  dtos.Post:
    properties:
      author:
        $ref: '#/definitions/dtos.UserInfo'
      comments:
        items:
          $ref: '#/definitions/dtos.Post'
        type: array
      comments_count:
        type: integer
      created_at:
        type: string
      likes_count:
        type: integer
      media:
        items:
          $ref: '#/definitions/dtos.PostMedia'
        type: array
      message:
        type: string
      post_id:
        type: string
      updated_at:
        type: string
    type: object
  dtos.PostMedia:
    properties:
      alternate_text:
        type: string
      url:
        type: string
    type: object
...
  dtos.UserInfo:
    properties:
      business_name:
        type: string
      first_name:
        type: string
      is_verified:
        type: boolean
      last_name:
        type: string
      profile_image:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      developer_information:
        example: an error occurred
        type: string
      error:
        example: message
        type: string
      status:
        example: error
        type: string
    type: object
...
  posts.createCommentResponse:
    properties:
      post:
        $ref: '#/definitions/dtos.Post'
      status:
        type: string
    type: object
  posts.createPostResponse:
    properties:
      post:
        $ref: '#/definitions/dtos.Post'
      status:
        type: string
    type: object
  posts.deletePostResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  posts.getPostsResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.PagedPosts'
      status:
        type: string
    type: object
  posts.getTimelineResponse:
    properties:
      data:
        $ref: '#/definitions/dtos.PagedContent'
      status:
        type: string
    type: object
  posts.reportPostResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8083
info:
  contact: {}
  description: Celebut Backend REST endpoints
  title: Celebut API
  version: "1.0"
paths:
...
  /comments/{postID}:
    get:
      consumes:
      - application/json
      description: Get Post Comments
      operationId: get-post-comments
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.getPostsResponse'
      security:
      - Bearer: []
      summary: Get Post Comments
      tags:
      - Posts
    post:
      consumes:
      - multipart/form-data
      description: Add comment to a post
      operationId: create-comment
      parameters:
      - description: post identifier
        in: path
        name: postID
        required: true
        type: string
      - description: comments media
        in: formData
        items:
          type: file
        name: media
        type: array
      - example: What is happening?
        in: formData
        name: message
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.createCommentResponse'
      security:
      - Bearer: []
      summary: Create post comment
      tags:
      - Posts
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get session user's posts
      operationId: get-posts
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.getPostsResponse'
      security:
      - Bearer: []
      summary: Get Posts
      tags:
      - Posts
    post:
      consumes:
      - multipart/form-data
      description: Create a post
      operationId: create-post
      parameters:
      - description: posts media
        in: formData
        items:
          type: file
        name: media
        type: array
      - example: What is happening?
        in: formData
        name: message
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.createPostResponse'
      security:
      - Bearer: []
      summary: Create Post
      tags:
      - Posts
  /posts/{postID}:
    delete:
      consumes:
      - application/json
      description: Delete a post
      operationId: delete-post
      parameters:
      - description: post identifier
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.deletePostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete Post
      tags:
      - Posts
  /posts/{userID}:
    get:
      consumes:
      - application/json
      description: Get user's posts by user ID
      operationId: get-user-posts
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - description: user identifier
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.getPostsResponse'
      security:
      - Bearer: []
      summary: Get User Posts by ID
      tags:
      - Posts
  /posts/report/{postID}:
    post:
      consumes:
      - application/json
      description: Report a post
      operationId: report-post
      parameters:
      - description: post identifier
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.reportPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Report post
      tags:
      - Posts
...
  /timeline:
    get:
      consumes:
      - application/json
      description: Get session user's timeline
      operationId: get-timeline
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.getTimelineResponse'
      security:
      - Bearer: []
      summary: Get Timeline
      tags:
      - Posts
...
securityDefinitions:
  Auth-Token:
    description: Type token for unauthenticated requests
    in: header
    name: x-auth-token
    type: apiKey
  Bearer:
    description: '`Bearer token` for authenticated requests'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"